pyproject.toml
[tool.poetry]
name = "faceagepredict"
version = "0.1.0"
description = ""
authors = ["denimbluu <denim.bluu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.10.0.82"
pathlib = "^1.0.1"
pillow = "^10.3.0"
torch = "^2.3.1"
torchvision = "^0.18.1"
pretrainedmodels = "^0.7.4"
imgaug = "^0.4.0"
yacs = "^0.1.8"
tensorboard = "^2.17.0"
tqdm = "^4.66.4"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
psutil = "^5.9.8"
fastapi = "^0.111.0"
streamlit = "^1.35.0"
ffmpeg-python = "^0.2.0"


[tool.poetry.group.dev.dependencies]
ptpython = "^3.0.27"
mypy = "^1.10.0"
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
